ReadME.md
This is the repository for BS32011 Practical Project A1.  The aim of this project is to expand on my Python, MySQL and HTML skills, by creating a working parser in order to generate functioning MySQL tables in a query able database. This is then to be hosted in a HTML interface, allowing others to query the dataset off location. 

This repository includes:
                                   -My Parser, written in Kate, the parser is designed to open NCBI GEO full soft datasets, sorting the data into                               .txt file depending on the probes used and headers searched for. See the documentation in the file for further details. 

                                   -MySQL tables, these come in two formats, the tables I have created, as well as a file with instruction on how to generate your own database tables. See the files for further information.
 
                                  -HTML code to create a hostable web interface, with a simple search function, allowing users to query the database outside of the local host.


Overall this project is not yet complete, however in the future I hope to create a full functioning database, with a larger gene pool and more indepth classes than that currently used.


This is the link to the GEO dataset used in this project.    http://www.ncbi.nlm.nih.gov/sites/GDSbrowser?acc=GDS5093​


27/02/15