Create Website foundations
This is the basis for a working website that will allow for querying, however the pointer file is not currently working. Replace all variables with the variables you want for your own website. See http://www.w3schools.com/tags/att_body_bgcolor.asp for a better idea of waht can be done in HTML.